requirements.txt
PyYAML==3.12
certifi==2017.4.17
chardet==3.0.4
datadog==0.16.0
decorator==4.1.1
idna==2.5
requests==2.18.1
simplejson==3.11.1
urllib3==1.21.1